requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-cache-url==2.0.0
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.1.2
djangorestframework==3.6.4
html5lib==0.999999999
idna==2.6
olefile==0.44
Pillow==4.3.0
pytz==2017.3
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
wagtail==1.13
wagtailfontawesome==1.1.0
webencodings==0.5.1
Willow==1.0
